+++ apps/src/bin/anoma-node/cli.rs
@@ -1,7 +1,7 @@
 //! The docstrings on types and their fields with `derive(Clap)` are displayed
 //! in the CLI `--help`.
 use anoma::config::Config;
-use anoma::node::{gossip, shell};
+use anoma::node::{gossip, ledger};
 use anoma::{cli, config};
 use eyre::{Context, Result};
 
@@ -25,12 +25,12 @@ pub fn main() -> Result<()> {
         Some((cli::RUN_LEDGER_COMMAND, _)) => {
             let config = get_cfg(home);
             let ledger_cfg = config.ledger.unwrap_or_default();
-            shell::run(ledger_cfg).wrap_err("Failed to run Anoma node")
+            ledger::run(ledger_cfg).wrap_err("Failed to run Anoma node")
         }
         Some((cli::RESET_LEDGER_COMMAND, _)) => {
             let config = get_cfg(home);
             let ledger_cfg = config.ledger.unwrap_or_default();
-            shell::reset(ledger_cfg).wrap_err("Failed to reset Anoma node")
+            ledger::reset(ledger_cfg).wrap_err("Failed to reset Anoma node")
         }
         Some((cli::GENERATE_CONFIG_COMMAND, _args)) => {
             let gen_config = config::Config::generate(&home, false)
