+++ shared/src/ledger/mod.rs
@@ -1,3 +1,2 @@
 pub mod gas;
 pub mod storage;
-pub mod vm;
