+++ shared/src/ledger/storage/types.rs
