--- shared/src/protocol/vm/wasm/mod.rs
@@ -1,3 +1,0 @@
-pub mod memory;
-pub mod runner;
-pub mod wasm_host_env;
