+++ shared/src/vm/memory.rs
