+++ shared/src/vm/wasm/mod.rs
@@ -1,0 +1,3 @@
+pub mod wasm_host_env;
+pub mod wasm_memory;
+pub mod wasm_runner;
