+++ tests/src/vm_host_env/tx.rs
@@ -1,12 +1,12 @@
 use std::collections::HashSet;
 
-use anoma_shared::protocol::gas::BlockGasMeter;
-use anoma_shared::protocol::storage::mockdb::MockDB;
-use anoma_shared::protocol::storage::testing::TestStorage;
-use anoma_shared::protocol::vm;
-use anoma_shared::protocol::vm::prefix_iter::PrefixIterators;
-use anoma_shared::protocol::vm::write_log::WriteLog;
+use anoma_shared::ledger::gas::BlockGasMeter;
+use anoma_shared::ledger::storage::mockdb::MockDB;
+use anoma_shared::ledger::storage::testing::TestStorage;
+use anoma_shared::ledger::storage::write_log::WriteLog;
 use anoma_shared::types::{Address, Key};
+use anoma_shared::vm;
+use anoma_shared::vm::prefix_iter::PrefixIterators;
 
 /// This module combines the native host function implementations from
 /// [`native_tx_host_env`] above with the functions exposed to the tx wasm
@@ -73,9 +73,9 @@ mod native_tx_host_env {
 
     use std::cell::RefCell;
 
-    use anoma_shared::protocol::storage::testing::Sha256Hasher;
-    use anoma_shared::protocol::vm::host_env::*;
-    use anoma_shared::protocol::vm::memory::testing::NativeMemory;
+    use anoma_shared::ledger::storage::testing::Sha256Hasher;
+    use anoma_shared::vm::host_env::*;
+    use anoma_shared::vm::memory::testing::NativeMemory;
     // TODO replace with `std::concat_idents` once stabilized (https://github.com/rust-lang/rust/issues/29599)
     use concat_idents::concat_idents;
