+++ tests/src/vm_host_env/vp.rs
@@ -1,13 +1,13 @@
 use std::collections::HashSet;
 
-use anoma_shared::protocol::gas::VpGasMeter;
-use anoma_shared::protocol::storage::mockdb::MockDB;
-use anoma_shared::protocol::storage::testing::TestStorage;
-use anoma_shared::protocol::vm;
-use anoma_shared::protocol::vm::prefix_iter::PrefixIterators;
-use anoma_shared::protocol::vm::write_log::WriteLog;
+use anoma_shared::ledger::gas::VpGasMeter;
+use anoma_shared::ledger::storage::mockdb::MockDB;
+use anoma_shared::ledger::storage::testing::TestStorage;
+use anoma_shared::ledger::storage::write_log::WriteLog;
 use anoma_shared::types::address::{self, Address};
 use anoma_shared::types::Key;
+use anoma_shared::vm;
+use anoma_shared::vm::prefix_iter::PrefixIterators;
 
 /// This module combines the native host function implementations from
 /// [`native_vp_host_env`] above with the functions exposed to the vp wasm
@@ -84,9 +84,9 @@ mod native_vp_host_env {
 
     use std::cell::RefCell;
 
-    use anoma_shared::protocol::storage::testing::Sha256Hasher;
-    use anoma_shared::protocol::vm::host_env::*;
-    use anoma_shared::protocol::vm::memory::testing::NativeMemory;
+    use anoma_shared::ledger::storage::testing::Sha256Hasher;
+    use anoma_shared::vm::host_env::*;
+    use anoma_shared::vm::memory::testing::NativeMemory;
     // TODO replace with `std::concat_idents` once stabilized (https://github.com/rust-lang/rust/issues/29599)
     use concat_idents::concat_idents;
